Add files via upload
last version

+++ models/author.py
@@ -1,6 +1,5 @@
 # -*- coding: utf-8 -*-
-from odoo import api, fields, models
-
+from odoo import api, fields, models,exceptions
 
 class Author(models.Model):
 
@@ -8,7 +7,7 @@ class Author(models.Model):
     _inherits = {'res.partner': 'partner_id'}
     _order = 'name asc'
 
-    # Relation with Book Model
+    # Field
     is_book_author = fields.Boolean(
         'Is Book Author',
         required=True,
@@ -21,10 +20,10 @@ class Author(models.Model):
         string="Author"
     )
 
-    # If we delete author we should delete his books
     @api.multi
     def unlink(self):
-        books = self.env['about.book'].search([('author_ids', 'in', self.ids)])
-        if books:
-            books.unlink()
+        book = self.env['about.book']
+        book_ids = book.search([('author_ids', 'in', self.ids)]) 
+        if book_ids:
+            raise exceptions.Warning("You are trying to delete an Author who is still referenced!")
         return super(Author, self).unlink()
